requirements.txt
locust==2.31.7
paho-mqtt==2.1.0

### pip install -r requirements.txt